requirements.txt

# Core dependencies
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.10.0
google-api-python-client>=2.108.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Document processing
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
python-pptx>=0.6.23

# Data handling
pyyaml>=6.0.1
python-dateutil>=2.8.2
pytz>=2023.3

# Utilities
requests>=2.31.0
tenacity>=8.2.3
tqdm>=4.66.0

# Logging and monitoring
structlog>=23.2.0
python-json-logger>=2.0.7

# CLI
click>=8.1.7
rich>=13.7.0
prompt-toolkit>=3.0.41

# Web Server (For Cloud Run)
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
